+++ src/data/input_provider_test.ts
@@ -15,7 +15,8 @@
  * =============================================================================
  */
 
-import {NDArrayMathCPU, NDArrayMathGPU} from '../math/math';
+import {NDArrayMathCPU} from '../math/backends/backend_cpu';
+import {NDArrayMathGPU} from '../math/backends/backend_webgl';
 import {Array1D, Scalar} from '../math/ndarray';
 
 import {InCPUMemoryShuffledInputProviderBuilder} from './input_provider';
