+++ src/graph/ops/argmaxequals_test.ts
@@ -15,7 +15,7 @@
  * =============================================================================
  */
 
-import {NDArrayMathCPU} from '../../math/math';
+import {NDArrayMathCPU} from '../../math/backends/backend_cpu';
 import {Array1D} from '../../math/ndarray';
 import {Tensor} from '../graph';
 import {TensorArrayMap} from '../tensor_array_map';
