+++ src/graph/ops/concat3d_test.ts
@@ -15,8 +15,8 @@
  * =============================================================================
  */
 
+import {NDArrayMathCPU} from '../../math/backends/backend_cpu';
 import * as concat_util from '../../math/concat_util';
-import {NDArrayMathCPU} from '../../math/math';
 import {Array3D} from '../../math/ndarray';
 import {Tensor} from '../graph';
 import {TensorArrayMap} from '../tensor_array_map';
