+++ src/graph/ops/divide_test.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  * =============================================================================
  */
-import {NDArrayMathCPU} from '../../math/math';
+import {NDArrayMathCPU} from '../../math/backends/backend_cpu';
 import {Array1D, Scalar} from '../../math/ndarray';
 import * as test_util from '../../test_util';
 import {Tensor} from '../graph';
