+++ src/graph/ops/log_test.ts
@@ -15,7 +15,7 @@
  * =============================================================================
  */
 
-import {NDArrayMathCPU} from '../../math/math';
+import {NDArrayMathCPU} from '../../math/backends/backend_cpu';
 import {Array1D} from '../../math/ndarray';
 import * as test_util from '../../test_util';
 import {Tensor} from '../graph';
