+++ src/graph/optimizers/momentum_optimizer_test.ts
@@ -15,7 +15,7 @@
  * =============================================================================
  */
 import {InputProvider} from '../../data/input_provider';
-import {NDArrayMathCPU} from '../../math/math';
+import {NDArrayMathCPU} from '../../math/backends/backend_cpu';
 import {Array1D, NDArray} from '../../math/ndarray';
 import * as test_util from '../../test_util';
 import {Graph} from '../graph';
