+++ src/graph/tensor_array_map_test.ts
@@ -15,7 +15,7 @@
  * =============================================================================
  */
 
-import {NDArrayMathCPU} from '../math/math';
+import {NDArrayMathCPU} from '../math/backends/backend_cpu';
 import {Array1D, NDArray} from '../math/ndarray';
 
 import {Tensor} from './graph';
