+++ src/index.ts
@@ -46,10 +46,12 @@ export {CostReduction, FeedEntry, Session} from './graph/session';
 export {GraphRunner, GraphRunnerEventObserver, MetricReduction} from './graph_runner';
 // tslint:disable-next-line:max-line-length
 export {ConstantInitializer, Initializer, NDArrayInitializer, OnesInitializer, RandomNormalInitializer, RandomTruncatedNormalInitializer, RandomUniformInitializer, VarianceScalingInitializer, ZerosInitializer} from './initializers';
-export {MatrixOrientation} from './math/backends/math_backend';
+export {MatrixOrientation} from './math/backends/backend';
+export {NDArrayMathCPU} from './math/backends/backend_cpu';
+export {NDArrayMathGPU} from './math/backends/backend_webgl';
 export {GPGPUContext} from './math/backends/webgl/gpgpu_context';
 // tslint:disable-next-line:max-line-length
-export {LSTMCell, NDArrayMath, NDArrayMathCPU, NDArrayMathGPU} from './math/math';
+export {LSTMCell, NDArrayMath} from './math/math';
 // TODO(nsthorat): Remove this once we decouple NDArray from storage mechanism.
 export {initializeGPU} from './math/ndarray';
 // tslint:disable-next-line:max-line-length
