+++ src/math/cost_functions_test.ts
@@ -15,9 +15,10 @@
  * =============================================================================
  */
 
+import {NDArrayMathCPU} from '../math/backends/backend_cpu';
 import * as test_util from '../test_util';
+
 import {SquareCostFunc} from './cost_functions';
-import {NDArrayMathCPU} from './math';
 import {Array1D} from './ndarray';
 
 describe('Cost functions', () => {
