+++ src/math/onehot_test.ts
@@ -15,9 +15,11 @@
  * =============================================================================
  */
 
+import {NDArrayMathCPU} from '../math/backends/backend_cpu';
+import {NDArrayMathGPU} from '../math/backends/backend_webgl';
 import * as test_util from '../test_util';
+
 import {NDArrayMath} from './math';
-import {NDArrayMathCPU, NDArrayMathGPU} from './math';
 import {Array1D} from './ndarray';
 
 function executeTests(mathFactory: () => NDArrayMath) {
